Root effects won't flag the target as "stunned"
Replaced 0x40000 with UnitFlags::Stunned for readability
Fixed creature ai getting stuck on stun

+++ tests/proto_patcher/main.cpp
@@ -1087,7 +1087,7 @@ namespace wowpp
 					continue;
 				}
 
-				if (spell->effects_size() <= effect)
+				if (UInt32(spell->effects_size()) <= effect)
 				{
 					WLOG("Spell " << spell->id() << " has only " << spell->effects_size() << " effects, but effect #" << effect << " was requested");
 					if (spell->effects_size() > 0)
